requirements.txt
fastapi[standard]==0.116.1
pyjwt==2.10.1
passlib==1.7.4
sqlmodel==0.0.24
pydantic_settings==2.10.1
# psycopg2==2.9.10
psycopg2-binary==2.9.10
bcrypt==4.0.1
alembic==1.16.4
python-box==7.3.2